+++ index.html
@@ -2,10 +2,22 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+    <meta name="description" content="{description}" />
+    <meta name="author" content="{author}" />
+    <!-- Twitter -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="{twitterCreator}" />
+    <meta name="twitter:creator" content="{twitterCreator}" />
+    <!-- OpenGraph (Facebook) -->
+    <meta property="og:title" content="{title}" />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="{url}" />
+    <meta property="og:image" content="{image}" />
+    <meta property="og:description" content="{description}" />
+    <meta name="theme-color" content="#FCBA03">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0">
     <link rel="icon" type="image/png" href="/icon.png" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Brella Counter</title>
-    <!--app-head-->
+    <title>{title}</title>
   </head>
   <body>
     <div id="root"><!--app-html--></div>
